chore: move InstanceContact and InstanceThumbnail to upper level
