+++ core/api/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/core/api/dto/InstanceContact.kt
@@ -1,0 +1,10 @@
+package com.livefast.eattrash.raccoonforfriendica.core.api.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class InstanceContact(
+    @SerialName("email") val email: String? = null,
+    @SerialName("account") val account: Account? = null,
+)
