+++ core/api/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/core/api/dto/InstanceThumbnail.kt
@@ -1,0 +1,11 @@
+package com.livefast.eattrash.raccoonforfriendica.core.api.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class InstanceThumbnail(
+    @SerialName("url") val url: String? = null,
+    @SerialName("blurhash") val blurhash: String? = null,
+    @SerialName("description") val description: String? = null,
+)
